+++ src/core/index.ts
@@ -5,8 +5,8 @@ import SchemaConvertor from './schemaConvertor';
 import WriteProcessor, { WriteProcessorOptions } from './writeProcessor';
 
 export interface Options extends Partial<WriteProcessorOptions> {
-    contents: any[];
-    inputUrls: string[];
+    contents?: any[];
+    inputUrls?: string[];
     header?: string;
 }
 
@@ -15,10 +15,14 @@ export default async function dtsGenerator(options: Options): Promise<string> {
     const resolver = new ReferenceResolver();
     const convertor = new SchemaConvertor(processor, options.header);
 
-    options.contents
-        .map((content) => parseSchema(content))
-        .forEach(resolver.registerSchema);
-    await Promise.all(options.inputUrls.map(resolver.registerRemoteSchema));
+    if (options.contents != null) {
+        options.contents
+            .map((content) => parseSchema(content))
+            .forEach(resolver.registerSchema);
+    }
+    if (options.inputUrls != null) {
+        await Promise.all(options.inputUrls.map(resolver.registerRemoteSchema));
+    }
 
     const generator = new DtsGenerator(resolver, convertor);
     return await generator.generate();
