+++ test/error_case_test.ts
@@ -1,5 +1,5 @@
 import assert from 'power-assert';
-import dtsgenerator from '../src/';
+import dtsgenerator from '../src/core';
 
 
 describe('error schema test', () => {
@@ -9,7 +9,7 @@ describe('error schema test', () => {
             type: 'object',
         };
         try {
-            await dtsgenerator([schema]);
+            await dtsgenerator({ contents: [schema] });
             assert.fail();
         } catch (e) {
             assert.equal(e.message, 'There is no id in the input schema(s)');
@@ -21,7 +21,7 @@ describe('error schema test', () => {
             type: 'hoge',
         };
         try {
-            await dtsgenerator([schema]);
+            await dtsgenerator({ contents: [schema] });
             assert.fail();
         } catch (e) {
             assert.equal(e.message, 'unknown type: hoge');
@@ -38,7 +38,7 @@ describe('error schema test', () => {
             },
         };
         try {
-            await dtsgenerator([schema]);
+            await dtsgenerator({ contents: [schema] });
             assert.fail();
         } catch (e) {
             assert.equal(e.message, 'unknown type: fuga');
@@ -56,7 +56,7 @@ describe('error schema test', () => {
             },
         };
         try {
-            await dtsgenerator([schema]);
+            await dtsgenerator({ contents: [schema] });
             assert.fail();
         } catch (e) {
             assert.equal(e.message, '$ref target is not found: /notFound/id#');
@@ -73,7 +73,7 @@ describe('error schema test', () => {
             },
         };
         try {
-            await dtsgenerator([schema]);
+            await dtsgenerator({ contents: [schema] });
             assert.fail();
         } catch (e) {
             assert.equal(e.message, '$ref target is not found: /test/target_not_found#hogefuga');
@@ -82,7 +82,7 @@ describe('error schema test', () => {
     it('invalid format schema', async () => {
         const schema = 'not schema data and invalid JSON format {.' as any;
         try {
-            await dtsgenerator([schema]);
+            await dtsgenerator({ contents: [schema] });
             assert.fail();
         } catch (e) {
             assert.ok(/^Unexpected token/.test(e.message));
